format code, 80 char margin

+++ src/ConvolveCPU.hs
@@ -1,23 +1,20 @@
 module ConvolveCPU where
-import qualified Data.Array.Repa as R
-import qualified Data.Array.Repa.Repr.Unboxed as RU
-import qualified Data.Array.Repa.Index as RI
-import qualified Data.Array.Repa.Algorithms.Convolve as RC
-import qualified Utils as U
+
+import qualified Data.Array.Repa                                as R
+import qualified Data.Array.Repa.Repr.Unboxed                   as RU
+import qualified Data.Array.Repa.Index                          as RI
+import qualified Data.Array.Repa.Algorithms.Convolve            as RC
+import qualified Utils                                          as U
 
 conv4D
-  -- :: (Num a, RU.Unbox a)
-  -- => 
   :: (Monad m) 
-  => RU.Array R.D RI.DIM4 Double
-  -> RU.Array R.D RI.DIM4 Double
+  => RU.Array R.D RI.DIM4 Float
+  -> RU.Array R.D RI.DIM4 Float
   -> Int
-  -> m(RU.Array R.D RI.DIM4 Double)
+  -> m(RU.Array R.D RI.DIM4 Float)
 conv4D img fltr strd = do
-  -- Neural network convolution
-  -- both inputs are 4d tensors, second dimension must match
-
-  -- convenice value - equal 1 + (imgRows- fltRows) / strd
+  -- Inputs: both array inputs are 4d tensors, second dimension must match,
+  --         int strd is a stride for the convultion
   -- Output: Delayed 4D tensor
   let imgDim = reverse $ U.los (R.extent img)
       fltrDim = reverse $ U.los (R.extent fltr)
@@ -25,13 +22,17 @@ conv4D img fltr strd = do
       bRange = [0..(imgDim !! 0)-1]
       kRange = [0..(fltrDim !! 0)-1]
       combRange = [(b,k) | b <- bRange, k <- kRange] 
-      mapHelper :: (Monad m) => (Int, Int) -> m(RU.Array RU.U RI.DIM2 Double)
+      mapHelper :: (Monad m) => (Int, Int) -> m(RU.Array RU.U RI.DIM2 Float)
       mapHelper (b,k) = do
         -- Takes the Image batchSize index and the filter batchSize index
         -- returns a 2d matrix as the resul of convolving using stride strd
         -- img[b, i, : , :] with fltr[k, i, :, :] for all i, and summing over i
         let iRange = [0..(imgDim !! 1)-1]
-            iResultsM = map conv2D [((R.slice img (R.Z R.:. (b :: Int) R.:. (i :: Int) R.:. R.All R.:. R.All)), (R.slice fltr (R.Z R.:. (k :: Int) R.:. (i :: Int) R.:. R.All R.:. R.All)), strd) | i <- iRange] 
+            iResultsM = map conv2D [((R.slice img (R.Z R.:. (b :: Int)
+                                      R.:. (i :: Int) R.:. R.All R.:. R.All)),
+                                     (R.slice fltr (R.Z R.:. (k :: Int) 
+                                      R.:. (i :: Int) R.:. R.All R.:. R.All)),
+                                     strd) | i <- iRange] 
         iResults <- U.unWrapList iResultsM
         sumOfRes <- R.computeUnboxedP (foldl (R.+^) (head iResults)
                       (tail iResults))
@@ -52,8 +53,8 @@ conv4D img fltr strd = do
 
 conv2D
   :: (Monad m)
-  => (RU.Array R.D RI.DIM2 Double, RU.Array R.D RI.DIM2 Double, Int)
-  -> m (RU.Array R.D RI.DIM2 Double)
+  => (RU.Array R.D RI.DIM2 Float, RU.Array R.D RI.DIM2 Float, Int)
+  -> m (RU.Array R.D RI.DIM2 Float)
 conv2D (img, fltr, strd) = do
   -- vanilla 2d convultion with stride strd - very hackish fuction
